modularize table of contents
